fix: move MnemonicStrengthType to shared/types/wallet.ts
