+++ shared/types/wallet.ts
@@ -80,3 +80,5 @@ export interface WalletFile {
   meta: WalletMeta;
   crypto: WalletSecretsEncrypted;
 }
+
+export type MnemonicStrengthType = 12 | 24;
